pyproject.toml
[project]
name = "peerreview-be"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "ipfshttpclient>=0.7.0",
    "py-solc-x>=2.0.4",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.41",
    "web3>=7.12.0",
]
